requirements.txt
databento>=0.64.0
matplotlib>=3.9.0
python-dotenv>=1.1.0
pandas>=2.3.0
numpy>=2.0.0
python-dateutil>=2.9.0
